% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_bw.R
\name{boot_bw_weight}
\alias{boot_bw_weight}
\alias{boot_bw_sample_cluster}
\alias{boot_bw_sample_clusters}
\alias{boot_bw_sample_within_cluster}
\alias{boot_bw_sample_within_clusters}
\alias{boot_bw}
\title{Blocked weighted bootstrap - vectorised and parallel}
\usage{
boot_bw_weight(w)

boot_bw_sample_cluster(p, w)

boot_bw_sample_clusters(x, w, index = FALSE)

boot_bw_sample_within_cluster(cluster_df)

boot_bw_sample_within_clusters(cluster_df)

boot_bw(
  x,
  w,
  statistic,
  params,
  outputColumns = params,
  replicates = 399,
  cores = parallelly::availableCores(omit = 1)
)
}
\arguments{
\item{w}{A data frame with primary sampling unit (PSU) in column named \code{psu}
and survey weight (i.e. PSU population) in column named \code{pop}}

\item{p}{A random probability of selection}

\item{x}{A data frame with primary sampling unit (PSU) in column named \code{psu}}

\item{index}{Logical. Should index values be returned or a list of
data.frames. Default to FALSE.}

\item{cluster_df}{A \code{\link[=data.frame]{data.frame()}} or a list of \code{\link[=data.frame]{data.frame()}}s for selected
clusters}

\item{statistic}{A function operating on data in \code{x} (see example)}

\item{params}{Parameters (named columns in \code{x}) passed to the function
specified in \code{statistic}}

\item{outputColumns}{Names of columns in output data frame}

\item{replicates}{Number of bootstrap replicates}

\item{cores}{The number of computer cores to use/number of child processes
will be run simultaneously.}
}
\value{
A \code{\link[=data.frame]{data.frame()}} with:
\itemize{
\item ncol = length(outputColumns)
\item nrow = replicates
\item names = outputColumns
}
}
\description{
Blocked weighted bootstrap - vectorised and parallel
}
\examples{
boot_bw(
  x = indicatorsHH, w = villageData, statistic = bootClassic, 
  params = "anc1", replicates = 49
)

}
