% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootClassic.R
\name{bootClassic}
\alias{bootClassic}
\title{bootClassic}
\usage{
bootClassic(x, params)
}
\arguments{
\item{x}{A data frame with \code{primary sampling unit (PSU)} in column named
\code{psu} and with data column/s containing the binary variable/s (0/1) of
interest with column names corresponding to \code{params} values}

\item{params}{A vector of column names corresponding to the binary variables
of interest contained in \code{x}}
}
\value{
A numeric vector of the mean of each binary variable of interest with
  length equal to \code{length(params)}
}
\description{
Simple proportion statistics function for bootstrap estimation
}
\examples{

# Example call to bootClassic function

meanResults <- bootClassic(x = indicatorsHH,
                           params = "anc1")

}
