% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_strat_estimation.R
\name{calc_total_ci}
\alias{calc_total_ci}
\title{Calculate confidence limits}
\usage{
calc_total_ci(est, pop, se, ci = c("lcl", "ucl"))
}
\description{
Calculate confidence limits
}
\keyword{internal}
