% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_strat_estimation.R
\name{estimate_total_}
\alias{estimate_total_}
\title{Estimate post-stratification weighted totals}
\usage{
estimate_total_(est_pop_df)
}
\description{
Estimate post-stratification weighted totals
}
\keyword{internal}
