% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{indicatorsCH1}
\alias{indicatorsCH1}
\title{Child Morbidity, Health Service Coverage, Anthropometry}
\format{
A data frame with 16 columns and 3090 rows.\tabular{ll}{
   \strong{Variable} \tab \strong{Description} \cr
   \code{region} \tab Region in Somalia from which the cluster belongs to \cr
   \code{district} \tab District in Somalia from which the cluster belongs to \cr
   \code{psu} \tab The PSU identifier. This must use the same coding system used to identify the PSUs that is used in the indicators dataset \cr
   \code{mID} \tab The mother identifier \cr
   \code{cID} \tab The child identifier \cr
   \code{ch1} \tab Diarrhoea in the past 2 weeks (0/1) \cr
   \code{ch2} \tab Fever in the past 2 weeks (0/1) \cr
   \code{ch3} \tab Cough in the past 2 weeks (0/1) \cr
   \code{ch4} \tab Immunisation card (0/1) \cr
   \code{ch5} \tab BCG immunisation (0/1) \cr
   \code{ch6} \tab Vitamin A coverage in the past month (0/1) \cr
   \code{ch7} \tab Anti-helminth coverage in the past month (0/1) \cr
   \code{sex} \tab Sex of child \cr
   \code{muac1} \tab Mid-upper arm circumference in mm \cr
   \code{muac2} \tab Mid-upper arm circumference in mm \cr
   \code{oedema} \tab Oedema (0/1) \cr
}
}
\source{
Mother and child health and nutrition survey in 3 regions of Somalia
}
\usage{
indicatorsCH1
}
\description{
Child indicators on morbidity, health service coverage and anthropometry
calculated from survey data collected in survey conducted in 4 districts
from 3 regions in Somalia.
}
\examples{
indicatorsCH1

}
\keyword{datasets}
