% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{indicatorsCH2}
\alias{indicatorsCH2}
\title{Infant and Child Feeding Index}
\format{
A data frame with 15 columns and 2083 rows.\tabular{ll}{
   \strong{Variable} \tab \strong{Description} \cr
   \code{region} \tab Region in Somalia from which the cluster belongs to \cr
   \code{district} \tab District in Somalia from which the cluster belongs to \cr
   \code{psu} \tab The PSU identifier. This must use the same coding system used to identify the PSUs that is used in the indicators dataset \cr
   \code{mID} \tab The mother identifier \cr
   \code{cID} \tab The child identifier \cr
   \code{ebf} \tab Exclusive breastfeeding (0/1) \cr
   \code{cbf} \tab Continued breastfeeding (0/1) \cr
   \code{ddd} \tab Dietary diversity (0/1) \cr
   \code{mfd} \tab Meal frequency (0/1) \cr
   \code{icfi} \tab Infant and child feeding index (from 0 to 6) \cr
   \code{iycf} \tab Good IYCF \cr
   \code{icfiProp} \tab Good ICFI \cr
   \code{age} \tab Child's age \cr
   \code{bf} \tab Child is breastfeeding (0/1) \cr
   \code{bfStop} \tab Age in months child stopped breastfeeding \cr
}
}
\source{
Mother and child health and nutrition survey in 3 regions of Somalia
}
\usage{
indicatorsCH2
}
\description{
Infant and young child feeding indicators using the infant and child feeding
index (ICFI) by Arimond and Ruel. Calculated from survey data collected in
survey conducted in 4 districts from 3 regions in Somalia.
}
\examples{
indicatorsCH2

}
\keyword{datasets}
