% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{indicatorsHH}
\alias{indicatorsHH}
\title{Mother Indicators Dataset}
\format{
A data frame with 26 columns and 2136 rows:\tabular{ll}{
   \strong{Variable} \tab \strong{Description} \cr
   \code{region} \tab Region in Somalia from which the cluster belongs to \cr
   \code{district} \tab District in Somalia from which the cluster belongs to \cr
   \code{psu} \tab The PSU identifier. This must use the same coding system used to identify the PSUs that is used in the indicators dataset \cr
   \code{mID} \tab The mother identifier \cr
   \code{mMUAC} \tab Mothers with mid-upper arm circumference < 230 mm (0/1) \cr
   \code{anc1} \tab At least 1 antenatal care visit with a trained health professional (0/1) \cr
   \code{anc2} \tab At least 4 antenatal care visits with any service provider (0/1) \cr
   \code{anc3} \tab FeFol coverage (0/1) \cr
   \code{anc4} \tab Vitamin A coverage (0/1) \cr
   \code{wash1} \tab Improved sources of drinking water (0/1) \cr
   \code{wash2} \tab Improved sources of other water (0/1) \cr
   \code{wash3} \tab Probable safe drinking water (0/1) \cr
   \code{wash4} \tab Number of litres of water collected in a day \cr
   \code{wash5} \tab Improved toilet facilities (0/1) \cr
   \code{wash6} \tab Human waste disposal practices / behaviour (0/1) \cr
   \code{wash7a} \tab Handwashing score (from 0 to 5) \cr
   \code{wash7b} \tab Handwashing score of 5 (0/1) \cr
   \code{hhs1} \tab Household hunger score (from 0 to 6) \cr
   \code{hhs2} \tab Little or no hunger (0/1) \cr
   \code{hhs3} \tab Moderate hunger (0/1) \cr
   \code{hhs4} \tab Severe hunger (0/1) \cr
   \code{mfg} \tab Mother's dietary diversity score \cr
   \code{pVitA} \tab Plant-based vitamin A-rich foods (0/1) \cr
   \code{aVitA} \tab Animal-based vitamin A-rich foods (0/1) \cr
   \code{xVitA} \tab Any vitamin A-rich foods (0/1) \cr
   \code{iron} \tab Iron-rich foods (0/1) \cr
}
}
\source{
Mother and child health and nutrition survey in 3 regions of Somalia
}
\usage{
indicatorsHH
}
\description{
Mother indicators for health and nutrition calculated from survey data
collected in survey conducted in 4 districts from 3 regions in Somalia.
}
\examples{
indicatorsHH

}
\keyword{datasets}
