% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{recode}
\alias{recode}
\title{Recode function}
\usage{
recode(var, recodes, afr, anr = TRUE, levels)
}
\arguments{
\item{var}{Variable to recode}

\item{recodes}{Character string of recode specifications:
  \itemize{
    \item Recode specifications in a character string separated by
    semicolons of the form input=output as in:

        "1=1;2=1;3:6=2;else=NA"

    \item If an input value satisfies more than one specification, then the
    first (reading from left to right) is applied

    \item If no specification is satisfied, then the input value is carried
    over to the result unchanged

    \item NA is allowed on both input and output

    \item The following recode specifications are supported :

          +---------------+------------------+---------------------------+
          | Specification | Example          | Notes                     |
          +---------------+------------------+---------------------------+
          | Single values | 9=NA             |                           |
          +---------------+------------------+---------------------------+
          | Set of values | c(1,2,5)=1       | The left-hand-side is any |
          |               | seq(1,9,2)='odd' | any R function call that  |
          |               | 1:10=1           | returns a vector          |
          +---------------+------------------+---------------------------+
          | Range of      | 7:9=3            | Special values lo and hi  |
          | or values     | lo:115=1         | may be used               |
          +---------------+------------------+---------------------------+
          | Other values  | else=NA          |                           |
          +---------------+------------------+---------------------------+

    \item Character values are quoted as in :

         recodes = "c(1,2,5)='sanitary' else='unsanitary'"

    \item The output may be the (scalar) result of a function call as in:

         recodes = "999=median(var, na.rm = TRUE)"

    \item Users are advised to carefully check the results of recode() calls
    with any outputs that are the results of a function call.

    \item The ouput may be the (scalar) value of a variable as in:

         recodes = "999=scalarVariable"

    \item If all of the output values are numeric, and if 'afr'' is FALSE,
    then a numeric result is returned; if var is a factor then
    (by default) so is the result.
}}

\item{afr}{Return a factor. Default is TRUE if \code{var} is a factor and is
FALSE otherwise}

\item{anr}{Coerce result to numeric (default is TRUE)}

\item{levels}{Order of the levels in the returned factor; the default is to use
the sort order of the level names.}
}
\value{
Recoded variale
}
\description{
Recode function
}
\examples{

#


}
