% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{somalia_population}
\alias{somalia_population}
\title{Somalia regional population in 2022}
\format{
An object of class \code{data.frame} with 19 rows and 18 columns.
}
\source{
\url{https://fsnau.org/downloads/2022-Gu-IPC-Population-Tables-Current.pdf}
}
\usage{
somalia_population
}
\description{
A data.frame with 19 rows and 18 columns:
}
\details{
\tabular{ll}{
   \strong{Variable} \tab \strong{Description} \cr
   \code{region} \tab Region name \cr
   \code{total} \tab Total population \cr
   \code{urban} \tab Total urban population \cr
   \code{rural} \tab Total rural population \cr
   \code{idp} \tab Total IDP population \cr
   \code{urban_stressed} \tab Total urban population - stressed \cr
   \code{rural_stressed} \tab Total rural population - stressed \cr
   \code{idp_stressed} \tab Total IDP population - stressed \cr
   \code{urban_crisis} \tab Total urban population - crisis \cr
   \code{rural_crisis} \tab Total rural population - crisis \cr
   \code{idp_crisis} \tab Total IDP population - crisis \cr
   \code{urban_emergency} \tab Total urban population - emergency \cr
   \code{rural_emergency} \tab Total rural population - emergency \cr
   \code{idp_emergency} \tab Total IDP population - emergency \cr
   \code{urban_catastrophe} \tab Total urban population - catastrophe \cr
   \code{rural_catastrophe} \tab Total rural population - catastrophe \cr
   \code{idp_catastrophe} \tab Total IDP population - catastrophe \cr
   \code{percent_at_least_crisis} \tab Percentage of population that are at least in crisis \cr
}
}
\keyword{datasets}
