% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_bw_estimate.R
\name{structure_estimates}
\alias{structure_estimates}
\title{Structure boot estimate outputs}
\usage{
structure_estimates(boot_df, est)
}
\description{
Structure boot estimate outputs
}
\keyword{internal}
