% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{indicatorsCH1}
\alias{indicatorsCH1}
\title{Child Morbidity, Health Service Coverage, Anthropometry}
\format{
A data frame with 14 columns and 3090 rows.
\describe{
\item{\code{psu}}{The PSU identifier. This must use the same coding system
used to identify the PSUs that is used in the indicators dataset}
\item{\code{mID}}{The mother identifier}
\item{\code{cID}}{The child identifier}
\item{\code{ch1}}{Diarrhoea in the past 2 weeks (0/1)}
\item{\code{ch2}}{Fever in the past 2 weeks (0/1)}
\item{\code{ch3}}{Cough in the past 2 weeks (0/1)}
\item{\code{ch4}}{Immunisation card (0/1)}
\item{\code{ch5}}{BCG immunisation (0/1)}
\item{\code{ch6}}{Vitamin A coverage in the past month (0/1)}
\item{\code{ch7}}{Anti-helminth coverage in the past month (0/1)}
\item{\code{sex}}{Sex of child}
\item{\code{muac1}}{Mid-upper arm circumference in mm}
\item{\code{muac2}}{Mid-upper arm circumference in mm}
\item{\code{oedema}}{Oedema (0/1)}
}
}
\source{
Mother and child health and nutrition survey in 3 regions of Somalia
}
\usage{
indicatorsCH1
}
\description{
Child indicators on morbidity, health service coverage and anthropometry
calculated from survey data collected in survey conducted in 4 districts
from 3 regions in Somalia.
}
\examples{
indicatorsCH1

}
\keyword{datasets}
