% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbw.R
\docType{data}
\name{indicatorsCH2}
\alias{indicatorsCH2}
\title{Infant and Child Feeding Index}
\format{A data frame with 13 columns and 2083 rows.
\describe{
\item{\code{psu}}{The PSU identifier. This must use the same coding system
  used to identify the PSUs that is used in the indicators dataset}
\item{\code{mID}}{The mother identifier}
\item{\code{cID}}{The child identifier}
\item{\code{ebf}}{Exclusive breastfeeding (0/1)}
\item{\code{cbf}}{Continued breastfeeding (0/1)}
\item{\code{ddd}}{Dietary diversity (0/1)}
\item{\code{mfd}}{Meal frequency (0/1)}
\item{\code{icfi}}{Infant and child feeding index (from 0 to 6)}
\item{\code{iycf}}{Good IYCF}
\item{\code{icfiProp}}{Good ICFI}
\item{\code{age}}{Child's age}
\item{\code{bf}}{Child is breastfeeding (0/1)}
\item{\code{bfStop}}{Age in months child stopped breastfeeding}
}}
\source{
Mother and child health and nutrition survey in 3 regions of Somalia
}
\usage{
indicatorsCH2
}
\description{
Infant and young child feeding indicators using the infant and child feeding
index (ICFI) by Arimond and Ruel. Calculated from survey data collected in
survey conducted in 4 districts from 3 regions in Somalia.
}
\keyword{datasets}
