% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbw.R
\docType{data}
\name{indicatorsHH}
\alias{indicatorsHH}
\title{Mother Indicators Dataset}
\format{A data frame with 24 columns and 2136 rows:
\describe{
\item{\code{psu}}{The PSU identifier. This must use the same coding system
  used to identify the PSUs that is used in the indicators dataset}
\item{\code{mID}}{The mother identifier}
\item{\code{mMUAC}}{Mothers with mid-upper arm circumference < 230 mm (0/1)}
\item{\code{anc1}}{At least 1 antenatal care visit with a trained health
  professional (0/1)}
\item{\code{anc2}}{At least 4 antenatal care visits with any service
  provider (0/1)}
\item{\code{anc3}}{FeFol coverage (0/1)}
\item{\code{anc4}}{Vitamin A coverage (0/1)}
\item{\code{wash1}}{Improved sources of drinking water (0/1)}
\item{\code{wash2}}{Improved sources of other water (0/1)}
\item{\code{wash3}}{Probable safe drinking water (0/1)}
\item{\code{wash4}}{Number of litres of water collected in a day}
\item{\code{wash5}}{Improved toilet facilities (0/1)}
\item{\code{wash6}}{Human waste disposal practices / behaviour (0/1)}
\item{\code{wash7a}}{Handwashing score (from 0 to 5)}
\item{\code{wash7b}}{Handwashing score of 5 (0/1)}
\item{\code{hhs1}}{Household hunger score (from 0 to 6)}
\item{\code{hhs2}}{Little or no hunger (0/1)}
\item{\code{hhs3}}{Moderate hunger (0/1)}
\item{\code{hhs4}}{Severe hunger (0/1)}
\item{\code{mfg}}{Mother's dietary diversity score}
\item{\code{pVitA}}{Plant-based vitamin A-rich foods (0/1)}
\item{\code{aVitA}}{Animal-based vitamin A-rich foods (0/1)}
\item{\code{xVitA}}{Any vitamin A-rich foods (0/1)}
\item{\code{iron}}{Iron-rich foods (0/1)}
}}
\source{
Mother and child health and nutrition survey in 3 regions of Somalia
}
\usage{
indicatorsHH
}
\description{
Mother indicators for health and nutrition calculated from survey data
collected in survey conducted in 4 districts from 3 regions in Somalia.
}
\keyword{datasets}
