% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{villageData}
\alias{villageData}
\title{Cluster Population Weights Dataset}
\format{
A data frame with 6 columns and 117 rows:\tabular{ll}{
   \strong{Variable} \tab \strong{Description} \cr
   \code{region} \tab Region in Somalia from which the cluster belongs to \cr
   \code{district} \tab District in Somalia from which the cluster belongs to \cr
   \code{psu} \tab The PSU identifier. This must use the same coding system used to identify the PSUs that is used in the indicators dataset \cr
   \code{lon} \tab Longitude coordinate of the cluster \cr
   \code{lat} \tab Latitude coordinate of the cluster \cr
   \code{pop} \tab Population size of the cluster \cr
}
}
\source{
Mother and child health and nutrition survey in 3 regions of Somalia
}
\usage{
villageData
}
\description{
Dataset containing cluster population weights for use in performing
posterior weighting with the blocked weighted bootstrap approach. This
dataset is from a mother and child health and nutrition survey conducted in
4 districts from 3 regions in Somalia.
}
\examples{
villageData

}
\keyword{datasets}
